+++ profiles/HeroAdvancement_Hero/Profile.lua
@@ -1,0 +1,1 @@
+HeroFreePick.RegisterMode('Hero')